fix update_axes
